# docs-site/docs/core/tournament-api.openapi.yaml
openapi: 3.0.3
info:
  title: Winspire Competition Tournament API
  version: 0.1.0
  description: |
    Player- and host-facing tournament management endpoints exposed by the
    Competition service (`services/competition`). All routes live under `/v1`
    and require a valid JWT when the service is configured with auth. Host
    routes additionally enforce host-admin authorization.
servers:
  - url: https://api.winspire.gg
    description: Production (example)
  - url: http://localhost:8080
    description: Local development (default Makefile target)
security:
  - bearerAuth: []
...
paths:
...
  /v1/{hostId}/tournaments/{tournamentId}:
    put:
      tags: [Tournament Host]
      operationId: editTournament
      summary: Edit a tournament
      description: |
        Updates mutable tournament fields (name, description, schedule, minimum
        team count). Requires host-admin access.
      parameters:
        - $ref: '#/components/parameters/HostIdParam'
        - $ref: '#/components/parameters/TournamentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditTournamentRequest'
      responses:
        '200':
          description: Tournament updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditTournamentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Tournament was not found under the given host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/{hostId}/tournaments/{tournamentId}/join:
    post:
      tags: [Tournament Player]
      operationId: joinTournament
      summary: Join a tournament
      description: Players join a tournament, optionally as a team or social team.
      parameters:
        - $ref: '#/components/parameters/HostIdParam'
        - $ref: '#/components/parameters/TournamentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinTournamentRequest'
      responses:
        '200':
          description: Join accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinTournamentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Tournament is closed or players already joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/{hostId}/tournaments/{tournamentId}/leave:
    post:
      tags: [Tournament Player]
      operationId: leaveTournament
      summary: Leave a tournament
      description: Removes the authenticated player/team from the tournament.
      parameters:
        - $ref: '#/components/parameters/HostIdParam'
        - $ref: '#/components/parameters/TournamentIdParam'
      responses:
        '200':
          description: Leave processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveTournamentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Participation record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/{hostId}/tournaments/{tournamentId}/confirm-participation:
    post:
      tags: [Tournament Player]
      operationId: confirmTournamentParticipation
      summary: Confirm tournament participation
      description: Confirms the player's participation once required by the host.
      parameters:
        - $ref: '#/components/parameters/HostIdParam'
        - $ref: '#/components/parameters/TournamentIdParam'
      responses:
        '200':
          description: Confirmation recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmTournamentParticipationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Tournament or participation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Confirmation window closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by the Host identity provider.
  parameters:
    HostIdParam:
      name: hostId
      in: path
      required: true
      description: UUID that identifies the host/workspace context.
      schema:
        type: string
        format: uuid
    TournamentIdParam:
      name: tournamentId
      in: path
      required: true
      description: UUID of the tournament within the host.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid path parameter or request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Caller lacks host-admin rights or tournament access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: Machine-readable error code.
        details:
          description: Additional error metadata or validation details.
      required: [error]
    UUID:
      type: string
      format: uuid
    DateTime:
      type: string
      format: date-time
      description: Timestamp encoded in RFC 3339 (UTC).
...
    EditTournamentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 4000
        status:
          type: string
          enum:
            - draft
            - open
            - scheduled
            - registration_open
            - started
            - completed
            - cancelled
          description: |
            Transition tournament to a new status:
            - `open`: Publish draft tournament (transitions to `scheduled` in database)
            - `registration_open`: Open registration for scheduled tournament
            - `started`: Start tournament (from registration_open or registration_closed)
            - `completed`: Complete tournament (from started)
            - `cancelled`: Cancel tournament (from any status except completed)
            
            **Note:** Backend validates status transitions and returns 400 for invalid transitions.
        scheduledStartTimeAt:
          $ref: '#/components/schemas/DateTime'
        minimumTeamCount:
          type: integer
          minimum: 1
    EditTournamentResponse:
      type: object
      properties:
        success:
          type: boolean
        errorCode:
          type: string
      required: [success]
...
    JoinTournamentRequest:
      type: object
      required: [players]
      properties:
        socialTeamId:
          $ref: '#/components/schemas/UUID'
        teamName:
          type: string
        players:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/UUID'
    JoinTournamentResponse:
      type: object
      properties:
        success:
          type: boolean
        errorCode:
          type: string
      required: [success]
    LeaveTournamentResponse:
      type: object
      properties:
        success:
          type: boolean
        errorCode:
          type: string
      required: [success]
    ConfirmTournamentParticipationResponse:
      type: object
      properties:
        success:
          type: boolean
        errorCode:
          type: string
      required: [success]
...